Code: updateComments.tsに移す
