+++ package.json
@@ -8,7 +8,7 @@
     "lint": "eslint --ext .js,.vue,.ts,.mts *.config.* src scripts",
     "fmt": "eslint --ext .js,.vue,.ts,.mts *.config.* src scripts --fix",
     "run:collect": "tsx scripts/collect.ts",
-    "run:update": "tsx scripts/update.ts",
+    "run:update": "tsx scripts/updateComments.ts",
     "dev": "vite",
     "build": "vue-tsc -b && vite build",
     "preview": "vite preview"
