+++ scripts/collect.ts
@@ -4,18 +4,15 @@ import { pipeline } from "node:stream/promises";
 import { Semaphore } from "@core/asyncutil";
 import unzip from "unzip-stream";
 import {
-  guestRepo,
   pagesBuildCheckName,
   artifactName,
   destinationDir,
-  pagesUrl,
-  commentMarker,
   rootLogger,
   octokit,
-  appInfo,
+  guestRepoOwner,
+  guestRepoName,
+  DownloadData,
 } from "./common.ts";
-
-const [guestRepoOwner, guestRepoName] = guestRepo.split("/");
 
 const branches = await octokit.paginate("GET /repos/{owner}/{repo}/branches", {
   owner: guestRepoOwner,
@@ -50,7 +47,7 @@ const downloadTargets = await Promise.all(
     ),
   ]
     .flat()
-    .map(async (source) => {
+    .map(async (source): Promise<DownloadData | undefined> => {
       const log = rootLogger.getChild(
         source.type === "branch"
           ? `Branch ${source.branch.name}`
@@ -193,65 +190,6 @@ const successfulDownloads = downloadTargets.filter(
 );
 if (successfulDownloads.length === 0) {
   throw new Error("No successful downloads");
-}
-
-for (const { dirname, source } of successfulDownloads) {
-  if (source.type === "branch") {
-    continue;
-  }
-  const log = rootLogger.getChild(`PR #${source.pullRequest.number}`);
-  log.info("Fetching comments...");
-  const comments = await octokit.paginate(
-    "GET /repos/{owner}/{repo}/issues/{issue_number}/comments",
-    {
-      owner: guestRepoOwner,
-      repo: guestRepoName,
-      issue_number: source.pullRequest.number,
-    },
-  );
-  const deployInfoMessage = [
-    ":rocket: プレビュー用ページを作成しました :rocket:",
-    "",
-    `- <a href="${pagesUrl}/preview/${dirname}/editor" target="_blank">:pencil: エディタ</a>`,
-    `- <a href="${pagesUrl}/preview/${dirname}/storybook" target="_blank">:book: Storybook</a>`,
-    "",
-    `更新時点でのコミットハッシュ：[\`${source.pullRequest.head.sha.slice(0, 7)}\`](https://github.com/${
-      source.pullRequest.head.repo.full_name
-    }/commit/${source.pullRequest.head.sha})`,
-    commentMarker,
-  ].join("\n");
-  const maybePreviousDeployInfo = comments.find(
-    (comment) =>
-      comment.user &&
-      appInfo.data &&
-      comment.user.login === `${appInfo.data.slug}[bot]` &&
-      comment.body?.endsWith(commentMarker),
-  );
-  if (!maybePreviousDeployInfo) {
-    log.info("Adding deploy info...");
-    await octokit.request(
-      "POST /repos/{owner}/{repo}/issues/{issue_number}/comments",
-      {
-        owner: guestRepoOwner,
-        repo: guestRepoName,
-        issue_number: source.pullRequest.number,
-        body: deployInfoMessage,
-      },
-    );
-  } else if (maybePreviousDeployInfo.body === deployInfoMessage) {
-    log.info("No update in deploy info, skipped.");
-  } else {
-    log.info("Updating deploy info...");
-    await octokit.request(
-      "PATCH /repos/{owner}/{repo}/issues/comments/{comment_id}",
-      {
-        owner: guestRepoOwner,
-        repo: guestRepoName,
-        comment_id: maybePreviousDeployInfo.id,
-        body: deployInfoMessage,
-      },
-    );
-  }
 }
 
 await fs.writeFile(
